+++ setup.py
@@ -10,7 +10,7 @@ setup(name='vosk-cli',
       author='Martin Wygas',
       author_email='mwygas@uos.de',
       license='Apache-2.0',
-      packages=['scripts'],
+      packages=['scripts', 'recasepunc'],
       license_files=('LICENSE'),
       include_package_data=True,
       zip_safe=False,
